+++ client/rtype/rtype/Powdered.hpp
@@ -19,7 +19,7 @@ public:
 	virtual void applyCollision(CollisionType type);
 	virtual void move(float delta);
 
-	void setColor(uint8_t color);
+	void setColor(uint16_t color);
 	bool isAnimationFinished(void) const;
 
 private:
@@ -29,7 +29,7 @@ private:
 	sf::RectangleShape *_shape;
 	float _delta;
 	uint8_t _currentFrame;
-	uint8_t _color;
+	uint16_t _color;
 	std::vector<sf::IntRect> _frames;
 	bool _finished;
 };
