+++ client/rtype/rtype/RequestDestroyed.cpp
@@ -13,7 +13,7 @@ void RequestDestroyed::execute(IClient &client, std::shared_ptr<ICommand> data,
 								std::shared_ptr<ICommand> &toSend)
 {
 	Destroyed *collision = (Destroyed *)data->getData();
-	uint8_t id = collision->id;
+	uint16_t id = collision->id;
 
 	StaticTools::Log << "executing destroyed request : id " << (int)id << std::endl;
 	AEntity *entity = World::getEntityByID(id);
